EVNT-144 Create Pricing Rules Layout and Connections (#276)
* Add Pricing Rules Page
Added scene and table for modifying pricing rules.

* Fix Up Some Styling

Re styled some components, check boxes now work appropriately oooo
aahhhhhh

* Use Redux to Manage Row State

Use redux to manage the state of the individual rows so changes to each
row can be tracked before saving the row.

* Updated Price Rule Changes to Satisfy Flow

Updated file changes related to the price rule table to be compliant
with flow.

* Add Tests for New Rule Pricing Table Components

Add tests for the components used by the price table scene as well as
the scene itself.

* Wire PriceScale, PriceRule Service, More Tests

Removed the sample json files being used as mock responses for price
rules and price scales and added calls to the respective APIS. Added
tests for the redux portion of the price rules, price scales and buyer
type additions.

Additionally, fixed a couple styling issues with the dropdowns on the
price rules table so the drop down would scroll and multi select would
be a little longer.

* Addressed Comments from Demo for Price Table UI
Updated several components on price rules table UI:
* price rules table link only available if tdc is the primary
* only one price rule can be updated at a time
* display selected values on multi select hover
* updated displayed values for rounding options
* other styling tweaks

* Update Price Rule Reducer, One Edit at a Time

Refactored the priceRule reducer to better reflect the constraint of
only allowing one row to be edited at a time. This commit removes
complexity around managing initial states for multiple price rows.

* Restructure Redux Store to Better Handle Edits

Updated the priceRule state reducer to focus on manipulate an ephemeral
editingRowState property instead of the priceRule directly. Also added a
hoc to wrap the state to prop and dispatch to props mapping outside of
the cell presenters. This abstracts interaction with redux out of the
presenter files themselves. Tests for the corresponding files were also
updated.