Update Price Rules Table Styling (#298)
Made some minor tweaks to clean up the pricing rules table. Updates
included:
* disabling outline when table is focused
* setting scene width to 100$ to align content width with header size
* adjusting column widths so enabled toggle does not get cut off
* rebuilt multi select so it has a similar style to the dropdown
* when editing, cancel and save buttons have room to breath at 1080 px
width
* multi select and drop down components will not have a height that
prevents the user from scrolling through all options, it will resize
based on available height
* updated number input component so plus and minus signs on ff and
chrome are allowed but letters are not

Additionally, moved the client integrations fetch so that it occurs
after authentication. It should now succeed regardless of the user's
initial state. As a result, the Pricing Rules option will be present in
the settings drop down for clients integrated with TDC.