Add Buyer Type Status Modal to Price Rule Scene (#299)
* Add Buyer Type Status Modal to Price Rule Scene

Added a button on the pricing rules page that will open up a modal with
a list of the buyer types, whether they are enabled and their
description. Additionally, there is an input at the top that will filter
the visible buyer types. Right now all buyer types are enabled, this
will be updated when the backend piece to pass the buyer type status has
been merged in.

* Move Filtering Input to Header, Add Tests

Move filtering input to the modal header so it is visible
while scrolling through buyer types. Added more tests for
the buyer type modal component.

* Add Disbaled to Buyer Type Model, Update Modal

Added the disabled property to the buyer type model. Added support for
displaying toggles on buyer type exlcusion modal that match values in
store. When the toggle values are changed, an updated copy of the buyer
type will be stored in the modals state such that it can be saved later
by clicking the save button on the modal.

Last but not least, added further tests for the buyer type exclusion
modal and the buyer type row component used by it.

* Lock Pricing Rules Scrolling When Modal Present

Updating pricing rules scene to hide overflow when the buyer type
exclusion modal is present. Additionally, add a header row to the
buyer type exclusion list.

* Factor Out Buyer Types Header Component

Moved the buyer type header out of the modal component into its own
function based component, updated styling to take better advantage of
flex sizing.

* Remove Logic From BuyerTypeRow Component

Move business logic to determine whether to display edited buyer type or
buyer type in store to the modal component consuming the BuyerTypeRow
component. Updated corresponding tests.