Add Ability to Enable and Disable Buyer Types (#308)
Connected the buyer types modal on the pricing rules scene to the API
that allows a user to enable or disable a specific buyer type. Right now
errors are handled very simply, by informing the user that the buyer
type could not be updated due a request issue or a conflict with an
existing pricing rule. Tests are included in this commit along with some
tweaks to the modal component logic.